+++ src/main/java/main/controller/DefaultController.java
@@ -1,13 +1,28 @@
 package main.controller;
 
 import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @Controller
 public class DefaultController {
-    @RequestMapping("/")
+    @RequestMapping(value = "/")
     public String index(){
         return "index";
     }
+    @GetMapping(value = "/**/{path:[^.]*}")
+    public String redirect(){
+        return "forward:/";
+    }
+
+//    @GetMapping(value = {
+//            "/{path:[^.]*}",
+//            "/{path:[^.]*}/{path:[^.]*}",
+//            "/{path:[^.]*}/{path:[^.]*}/{path:[^.]*}",
+//            "/{path:[^.]*}/{path:[^.]*}/{path:[^.]*}/{path:[^.]*}",
+//            "/{path:[^.]*}/{path:[^.]*}/{path:[^.]*}/{path:[^.]*}/{path:[^.]*}"
+//            // add more if required ...
+//    })
 }
